requirements.txt
asgiref==3.8.1
asttokens==2.4.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
crispy-bootstrap5==2024.10
cryptography==43.0.1
decorator==5.1.1
Django==5.1.1
django-allauth==65.0.2
django-crispy-forms==2.3
django-extensions==3.2.3
django-markdownx==4.0.7
executing==2.1.0
idna==3.10
ipython==8.28.0
jedi==0.19.1
Markdown==3.7
matplotlib-inline==0.1.7
parso==0.8.4
pillow==10.4.0
pip==23.2.1
prompt_toolkit==3.0.48
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
